Extend the axiomatic semantics for expressions.
It now includes a framing environment that is shared for function calls. It can
be used to keep track of invariants of shared data structures by all function
calls in an expression.

Besides, an equational theory for memories that behave equivalently with
respect to disjointness has been added. This make reasoning about disjoint
memories much easier, and is used to implement a faster and more reliable
tactic to automatically solve properties about disjoint memories involving
unions.