Remove conditions on size of memory objects.
Previously, there was a condition that each object in memory should have a size
that is at most max (ptrdiff_t). This restriction was too tight, in practice
objects may be bigger (for example, allocating a 3GB object on x86 actually
happens). This condition was solely used to show that pointer subtraction is
always representable by an integer of type ptrdiff_t. However, this is not the
case, see also 6.5.6p9 in the C11 standard. Hence, I removed it entirely.

This also makes some parts a lot cleaner.